requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pymongo==4.10.1
Pillow==11.0.0
graphviz==0.20.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.4